PatientDocumentAPI/requirements.txt
aniso8601==8.0.0
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cycler==0.10.0
FHIR-Parser==0.1.3
Flask==1.1.1
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
pycparser==2.19
pyparsing==2.4.6
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
